requirements.txt
streamlit
mysql-connector-python
os-sys